requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
comm==0.2.3
cryptography==46.0.2
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
nest-asyncio==1.6.0
openai==2.2.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyzmq==27.1.0
referencing==0.36.2
rich==14.1.0
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
validators==0.35.0
wcwidth==0.2.14
weaviate-agents==1.0.1
weaviate-client==4.17.0
fastapi==0.115.0
pyjwt==2.10.1
clerk-backend-api==1.7.0
